deps/libsmb2: fix -Wsign-compare warnings, one of them a real loop bug
The Windows build emits 35 -Wsign-compare warnings from the bundled
libsmb2. Most are cosmetic, but one is not, so they are worth clearing
rather than muting.

smb2-cmd-query-directory.c, both loops in
smb2_encode_query_directory_reply(), is an actual bug:

  in_remain = fslen;
  do {
          ...
          in_remain -= PAD_TO_64BIT(sizeof(struct smb2_fileidbothdirectoryinformation));
  } while (in_remain >= sizeof(struct smb2_fileidbothdirectoryinformation));

in_remain is int and sizeof is unsigned long long, so the comparison
converts in_remain to unsigned. The moment it goes negative - which it
does whenever fslen is not an exact multiple of the padded entry size -
it reads as a huge positive value, the condition stays true and the loop
keeps walking past the end of the buffer. Comparing against
(int)sizeof(...) makes it signed and the loop terminates.

The rest are type mismatches with no behaviour change:

  - libsmb2.c, sync.c: NTSTATUS values are 32-bit unsigned and the
    SMB2_STATUS_* constants above 0x7fffffff have unsigned int type, but
    smb2_command_cb carries status in an int. The comparisons were
    already being done in unsigned arithmetic by the usual conversions;
    the casts just say so. Changing the callback signature would be an
    API break across the whole library, so the casts stay at the
    comparison sites.
  - dcerpc.c, dcerpc-lsa.c: loop counters declared int while walking
    uint64_t/uint32_t counts that had already been used to size the
    matching allocation. Widened to match the count. For counts above
    INT_MAX the old code was signed overflow, so this is also a small
    correctness win.
  - ntlmssp.c: alloc_len held in an int and compared against uint32_t
    offsets. Made uint32_t, which is what the two assignments and both
    malloc() calls already treat it as.
  - smb2-data-file-info.c: name_len in
    smb2_encode_file_normalized_name_info() made uint32_t to match
    fs->file_name_length; the two comparisons in
    smb2_decode_file_stream_info() get explicit (size_t) casts spelling
    out the conversion that was already happening; the tv_usec sentinel
    check is cast to unsigned long, which preserves today's behaviour on
    both 32-bit and 64-bit long.
  - socket.c: smb2_accept_connection_async() takes an int fd but
    SMB2_INVALID_SOCKET is SOCKET-typed on Windows. Cast at the
    comparison. Worth noting separately that taking a Windows SOCKET
    through an int truncates on Win64 - that is a public prototype in
    libsmb2.h and is left alone here, and RetroArch does not use the
    accept path.

Deliberately not touched: the decode-side arithmetic in
smb2_decode_file_stream_info() still computes offsets in int from
network-supplied values. The casts silence the diagnostic and preserve
current behaviour, they do not harden the parser.

Verified with x86_64-w64-mingw32-gcc 13: all 55 sources in
Makefile.common's builtin libsmb2 object list go from 35 -Wsign-compare
warnings to 0, with 0 failures. A full Linux ./configure --enable-libsmb
rebuild links clean and produces an identically sized binary. Note the
53 -Wunknown-pragmas warnings my harness also showed are not present in
real builds - Makefile.common already passes -Wno-unknown-pragmas.

Worth sending upstream to libsmb2, the query-directory loop especially.

+++ deps/libsmb2/lib/dcerpc-lsa.c
@@ -91,7 +91,7 @@ lsa_RPC_SID_coder(struct dcerpc_context *dce,
 {
         RPC_SID *sid = ptr;
         uint64_t count;
-        int i;
+        uint64_t i;
 
         count = sid->SubAuthorityCount;
         if (dcerpc_uint3264_coder(dce, pdu, iov, offset, &count)) {
@@ -135,7 +135,7 @@ lsa_PRPC_SID_array_coder(struct dcerpc_context *dce,
 {
         PLSAPR_SID_ENUM_BUFFER seb = ptr;
         uint64_t val;
-        int i;
+        uint64_t i;
 
         val = seb->Entries;
         if (dcerpc_uint32_coder(dce, pdu, iov, offset, &val)) {
@@ -288,7 +288,7 @@ TN_array_coder(struct dcerpc_context *dce,
 {
         LSAPR_TRANSLATED_NAMES_EX *tn = ptr;
         uint64_t count;
-        int i;
+        uint64_t i;
 
         count = tn->Entries;
         if (dcerpc_uint3264_coder(dce, pdu, iov, offset, &count)) {
@@ -502,7 +502,7 @@ RDL_DOMAINS_array_coder(struct dcerpc_context *dce,
 {
         LSAPR_REFERENCED_DOMAIN_LIST *rdl = ptr;
         uint64_t entries;
-        int i;
+        uint64_t i;
 
         entries = rdl->Entries;
         if (dcerpc_uint3264_coder(dce, pdu, iov, offset, &entries)) {
